socat: version bump 1.7.3.2 -> 1.8.0.3
The patch is integrated in the new version. Also the ext2 config option
was moved to the more generic fs config option.

The README mentions that the project also uses MIT for the install-sh
file. As the configure script mentions that the file is required, the
license has been added to the list.

Signed-off-by: Sven Püschel <s.pueschel@pengutronix.de>
Message-Id: <20251014155656.1797756-1-s.pueschel@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>