fontconfig: version bump 2.14.2 -> 2.17.1
As noted on the fontconfig.org website, releases beyond 2.16.0 are
available on GitLab. Therefore change the URL accordingly.

While the GitLab releases contain an xz compressed tarball with a
generated configure script, they don't have a predictable path.
Therefore use the GitLab generated archives and switch to meson as a
build system to avoid adding an autogen.sh patch.

The CaseFolding license just bumped the year due to fontconfig including
a newer version of the file (14.0.0 -> 15.1.0).

https://gitlab.freedesktop.org/fontconfig/fontconfig/-/releases/2.16.1
https://gitlab.freedesktop.org/fontconfig/fontconfig/-/releases/2.16.2
https://gitlab.freedesktop.org/fontconfig/fontconfig/-/releases/2.17.0
https://gitlab.freedesktop.org/fontconfig/fontconfig/-/releases/2.17.1

Signed-off-by: Sven Püschel <s.pueschel@pengutronix.de>
Message-Id: <20251021081038.1373513-1-s.pueschel@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>