cryptsetup: version bump 2.7.4 -> 2.8.1
Cryptsetup has reorganized their licenses [1] and also added a general
readme about all licenses used in the project.

While the licenses in the README seems to match, it only mentions the
cryptsetup OpenSSL exception for the LGPL license and not the default
GPL-2.0-or-later license. This matches the SPDX Identifier in the source
files. But in contrast both included GPL-2.0-or-later licenses
(COPYING and COPYING.GPL-2.0-or-later-WITH-cryptsetup-OpenSSL-exception)
contain the OpenSSL exception clause. Therefore both licenses were
added to the LICENSE expression.

[1] https://git.kernel.org/pub/scm/utils/cryptsetup/cryptsetup.git/commit/?id=1a6a611dff6194071f6f374705e162500d656c53

Signed-off-by: Sven Püschel <s.pueschel@pengutronix.de>
Message-Id: <20251021083936.1420203-1-s.pueschel@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>