pam: version bump 1.5.2 -> 1.7.1
pam switched over to meson. Therefore adjust the options accordingly.
As the release download were only necessary for autoconfig, switch to
the automatically generated GitHub release archives.

With the switch to meson the libnsl dependency is now required when
enabling the NIS feature [1]. The library seemed to be optional in the
autoconf build [2]. Therefore mark the feature as BROKEN for now.

https://github.com/linux-pam/linux-pam/releases/tag/v1.5.3
https://github.com/linux-pam/linux-pam/releases/tag/v1.6.0
https://github.com/linux-pam/linux-pam/releases/tag/v1.6.1
https://github.com/linux-pam/linux-pam/releases/tag/v1.7.0
https://github.com/linux-pam/linux-pam/releases/tag/v1.7.1

[1] https://github.com/linux-pam/linux-pam/blob/c4a53492e1b7aebcf7a65a778d9e3a78f196d117/meson.build#L465
[2] https://github.com/linux-pam/linux-pam/blob/40f7d85f3736d058c26de1dafa4fed46de7d75ef/configure.ac#L441

Signed-off-by: Sven Püschel <s.pueschel@pengutronix.de>
Message-Id: <20251021091750.1548881-1-s.pueschel@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>