ptxd_make_get_git: add support for submodules
'git archive' has no support for submodules so this must be done
manually.

If submodules are enabled, then .gitmodules, for the relevant commit, is
parsed and sub-directory, URL and commit-ish are calculated. The
repositories are downloaded just like the main repository.

For each repository a tarball is created with a prefix that includes the
sub-directory. All tarball are concatenated and then compressed.
With the option '--ignore-zeros', tar can extract such archives
transparently.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>