xcb-proto: version bump 1.16.0 -> 1.17.0
The HTTP URL returns a 301 redirect. Therefore directly use the
redirect location as URL.

Also adjust the license to X11, as it contains a no advertising clause.
It doesn't quite match the X11 license, as it misses a trademark
statement. The SPDX Online tool [1] suggests
X11-distribute-modifications-variant, but as the COPYING file doesn't
contain the "distibute with modification" grant, using X11 alone is
the best choice.

[1] https://tools.spdx.org/app/check_license/

Signed-off-by: Sven Püschel <s.pueschel@pengutronix.de>
Message-Id: <20251022121443.200702-1-s.pueschel@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>