libxcb: version bump 1.16 -> 1.17.0
The HTTP URL redirects to the HTTPS URL. Therefore adjust the URL to
HTTPS to avoid the redirect.

Also fixing the SPDX license identifier, as the license text includes a
no advertising clause. While the SPDX Online tool [1] matches the
license text to X11-distribute-modifications-variant, it doesn't contain
the distribute with modifications grant. Therefore use the X11 license
which in contrast to MIT contains the advertising clause.
Note that it doesn't match exactly, as it misses the trademark at the
end, which caused the online tool to match the wrong license.

[1] https://tools.spdx.org/app/check_license/

Signed-off-by: Sven Püschel <s.pueschel@pengutronix.de>
Message-Id: <20251022121443.200702-2-s.pueschel@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>