libtag: version bump 1.13.1 -> 2.1.1
TagLib v2.0 is binary incompatible to v1.x, requiring all applications
to be rebuilt against it.

Formats can now be disabled at build time [1]. Keep the default by
keeping all formats enabled.

The UTF8 options are vendored from utf8cpp. While this header only
library can be installed into the system, it is not supported by the
author [2] and it should be vendored. Therefore don't create a separate
rule for utf8cpp and let taglib use it's vendored version.

The bindings option has been disabled, as it doesn't seem to be
required by gst-plugins-good1.

As the MPL license was included in the license files, but not the
license, the tag.h file header was inspected. It states that it
can either be licensed under the LGPL or the MPL and all other files
seem to share this OR relation. Also it doesn't mention the
"or later" phrase for the LGPL, therefore change the license to
LGPL-2.1-only.

https://mail.kde.org/pipermail/taglib-devel/2024-January/003075.html
https://mail.kde.org/pipermail/taglib-devel/2024-March/003079.html
https://mail.kde.org/pipermail/taglib-devel/2024-August/003081.html
https://mail.kde.org/pipermail/taglib-devel/2025-May/003096.html
https://mail.kde.org/pipermail/taglib-devel/2025-June/003103.html

[1] https://github.com/taglib/taglib/commit/ee1931b81116cd0091c906896f6f4fb74850be51
[2] https://github.com/nemtrif/utfcpp/blob/961e7a3e9b3c44c3f09a651694415030a4019bc2/README.md?plain=1#L121

Signed-off-by: Sven Püschel <s.pueschel@pengutronix.de>
Message-Id: <20251024080501.3278561-1-s.pueschel@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>