rtmpdump: version bump 2021-02-19-gf1b83c10d8be -> 2024-03-01-g6f6bb1353fc8
The double md5 sum was a result from a URL change without a new version.
Therefore two md5 sums were valid ones to keep the cached downloads
valid. As we now update the version, we don't have existing versions
in the cache and thereore can only use the md5 for the current URL.

Signed-off-by: Sven Püschel <s.pueschel@pengutronix.de>
Message-Id: <20251024215331.1696508-1-s.pueschel@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>