xkeyboard-config: version bump 2.42 -> 2.46
xkeyboard-config now uses a versioned directory to allow having
breaking change updates [1]. The /X11/xkb directory is now just a
symlink to the current version.

The xkb-base parameter wasn't used and therefore removed [2].
Instead the meson datadir was overwritten with the XORD_DATADIR to
match the variable usage in the targetinstall step.
Also the other meson options were added to the option list
with their default values.

The license file changed due to a license error fix [3].
As the COPYING file contains multiple licenses, each of them were
evaluated with the SPDX Online Tool [4] and added to the list of
licenses. The fixed license previously had the
xkeyboard-config-Zinoviev SPDX Identifier. As the fixed version doesn't
match any license and the SPDX identifier is very specific to
xkeyboard-config the old SPDX identifier is used.

The patch is not necessary anymore, as the 2.43 includes a fallback for
the Python 3.11 specific import [5].

[1] https://gitlab.freedesktop.org/xkeyboard-config/xkeyboard-config/-/commit/fd1d8d2d4f07ac494109b1a9e72d7fe777f6757a
[2] https://gitlab.freedesktop.org/xkeyboard-config/xkeyboard-config/-/commit/7f4bfd3af1b6448cc0e37a56a8f9a929c1219696
[3] https://gitlab.freedesktop.org/xkeyboard-config/xkeyboard-config/-/commit/8d42430f6498ceb1f9ca31f2fb9d0419c682912b
[4] https://tools.spdx.org/app/check_license/
[5] https://gitlab.freedesktop.org/xkeyboard-config/xkeyboard-config/-/commit/6993e6953deb5ce00647553b18e5273f15596b0b

Signed-off-by: Sven Püschel <s.pueschel@pengutronix.de>
Message-Id: <20251028084244.1985077-1-s.pueschel@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>