protobuf: version bump 3.7.1 -> 33.0
protobuf now uses bazel as it's primary build system, but also provides
a CMake build system. As PTXdist doesn't support bazel, use CMake.

Also protobuf now depends on the abseil-cpp library. But it doesn't
depend on the host protoc executable anymore, as it's only required
when tests or examples are built [1].

Compiling the protoc binary also needs libupb to be enabled, as the
compilation otherwise fails. Therefore it is enabled on the host side.

Protobuf now also depends on a UTF-8 Range library. This previously had
it's own repository, but is now developed inside of the protobuf
repository [2]. As the library is probably only used by protobuf and
has to keep in sync with the protobuf sources, avoid the effort of
separating it into a separate rule.

[1] https://github.com/protocolbuffers/protobuf/blob/a79f2d2e9fadd75e94f3fe40a0399bf0a5d90551/CMakeLists.txt#L317-L320
[2] https://github.com/protocolbuffers/utf8_range/blob/1d1ea7e3fedf482d4a12b473c1ed25fe0f371a45/README.md?plain=1#L1-L4

Signed-off-by: Sven Püschel <s.pueschel@pengutronix.de>
Message-Id: <20251028143921.512916-2-s.pueschel@pengutronix.de>
[mol: keep HOST_PROTOBUF so that users will find protoc]
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>