nbd: version bump 3.25 -> 3.26.1
The configure script from the tarball fails due to the man pages being
disabled, but docbook2man being installed. Therefore backport the
already upstreamed fix [1].

Due to the configure.ac file being changed by the patch, an autogen.sh
patch has to be added. As the nbd autogen.sh also generates a systemd
file, the autogen.sh from nbd was copied. This also brought up the
problem that a necessary file for the systemd generation process was
not included in the tarball. As the file was moved upstream, backport
the commit to have the file available and simplify a future upgrade by
simply dropping the whole patch folder.

https://github.com/NetworkBlockDevice/nbd/releases/tag/nbd-3.26
https://github.com/NetworkBlockDevice/nbd/releases/tag/nbd-3.26.1

[1] https://github.com/NetworkBlockDevice/nbd/pull/158

Signed-off-by: Sven Püschel <s.pueschel@pengutronix.de>
Message-Id: <20251028171252.1063580-1-s.pueschel@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>