host-imx-cst: change source origin to Debian snapshot
With the release of IMX-CST v4+ as only IMX_CST_TOOL_NEW variant, NXP no
longer provides any official download link for v3.4.1.

Switching to the latest release would require a lot of integration work
due to significant changes in the code base while not providing any
advantages for the platforms currently using v3.4.1.

We hence keep the former version for now and switch over to Debian
snapshot as source. This comes with the additional benefit of the
package now being aligned with the Debian Free Software Guidelines
(DFSG), which removes the need for accepting an EULA at the get stage.

Signed-off-by: Holger Assmann <h.assmann@pengutronix.de>
Message-Id: <20251103063320.30030-1-h.assmann@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>