report/sbom.py: ignore parameters for patches in series files
It is possible to have a parameter for patches in series
files like e.g. "-p0". This case will not work when parsing
the patches when generating the sbom, since the whole
string is taken to check in the path.exists check.

If the patch has parameters we skip those and only check for
the file which is always set first.

Signed-off-by: Michael Grzeschik <m.grzeschik@pengutronix.de>
Message-Id: <20251106112614.2342635-1-m.grzeschik@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>