ffmpeg: enable filters and swscale
Enable the filters and swscale, as they are required by the GStreamer
avvideocompare element. This uses the ssim or psnr filter to compare
two video inputs. Internally this also expects swscale to be available
to potentially scale the inputs to the same size.

The change increases the libavfilter.so file size from 148.0K to 4.3M
and adds the libswscale shared library with a size of 340K.
The remaining shared libraries from libav keep their size.

Signed-off-by: Sven Püschel <s.pueschel@pengutronix.de>
Message-Id: <20251128064720.2216811-1-s.pueschel@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>