scripts: split report generation
Currently, the license-report.yaml is generated after the targetinstall
stage. This is necessary because it includes the created ipkgs.

During development, this can be anoying, because "ptxdist licensecheck"
will build the packages.

Split the report into two stages: The existing license-report.yaml now
only depends on the extract stage, so "ptxdist licensecheck" is faster.
The new late-report.yaml contains just the parts that require a
completed package build. Currently, this is just the list of ipkgs.

The full-bsp-report explicitly includes both.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>