optee-client: use upstream udev rule and systemd template
Since optee-client version 4.4.0, the sources include a udev rule and a
systemd service file template in order to start tee-supplicant and
manage access to (OP-)TEE devices. Recent upstream changes have
streamlined this constellation and made it usable in a BSP without
additional setup required.

As these changes are not part of a proper release yet, we include them
into a patch stack for optee-client v4.8.0.

This commit removes the PTXdist downstream "tee-supplicant.service" in
favour of the upstream solution:

- The udev rule changes group ownership of every "/dev/tee*" device to
  "tee" and of every "/dev/treepriv*" as well as any eMMC RPMB partition
  to "teepriv". These group names are build time options and were already
  introduced to PTXdist with commit a06d32d59aa.
- The systemd service template derives instances for every  "/dev/teepriv*"
  device and starts a dedicated process of tee-supplicant. It is triggered
  by the aforementioned udev rule and executes as "teesuppl:teesuppl".

Signed-off-by: Holger Assmann <h.assmann@pengutronix.de>
Message-Id: <20251211102143.1411156-1-h.assmann@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>