kernel: handle building kernel rust code
The kernel config changes depending on whether rust support is detected.
So add an option to make this explicit and ensure that the support is
always or never detected accordingly.

'bindgen' is needed so select it and it must find the correct
llvm-config. The one from host-llvm (if built) does not work, so set
LLVM_CONFIG_PATH to explicitly skip it. That way, llvm-config from the
toolchain is used.

If the option is disabled, set RUSTC=false to ensure that rust support
is not detected.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>